Move jobs source to its own module.
